📝 chore: check if gettext is installed in build.sh

+++ build.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+gettext --version
+if [ $? -ne 0 ]; then
+    echo gettext is required to build the documentation.
+    exit 127
+fi
+
 rm -rf docs/*
 cp index.html docs/
